Refactor files out of server directory as it seems to be a special directory with unintended behavior
